# Conflicts:
#	discovery-service/src/main/java/org/zowe/apiml/discovery/GatewayNotifier.java
#	gateway-service/src/test/java/org/zowe/apiml/gateway/filters/pre/ZosmfFilterTest.java
#	gateway-service/src/test/java/org/zowe/apiml/gateway/security/login/zosmf/ZosmfAuthenticationProviderTest.java
#	gateway-service/src/test/java/org/zowe/apiml/gateway/security/query/SuccessfulQueryHandlerTest.java
#	gateway-service/src/test/java/org/zowe/apiml/gateway/security/service/AuthenticationServiceTest.java
#	gateway-service/src/test/java/org/zowe/apiml/gateway/security/service/schema/HttpBasicPassTicketSchemeTest.java

+++ discovery-service/src/main/java/org/zowe/apiml/discovery/GatewayNotifier.java
@@ -9,16 +9,19 @@
  */
 package org.zowe.apiml.discovery;
 
-import org.zowe.apiml.message.core.MessageService;
-import org.zowe.apiml.message.log.ApimlLogger;
-import org.zowe.apiml.util.EurekaUtils;
 import com.netflix.appinfo.InstanceInfo;
 import com.netflix.discovery.shared.Application;
 import com.netflix.eureka.EurekaServerContext;
 import com.netflix.eureka.EurekaServerContextHolder;
 import com.netflix.eureka.registry.PeerAwareInstanceRegistry;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
+import org.zowe.apiml.message.core.MessageService;
+import org.zowe.apiml.message.log.ApimlLogger;
+import org.zowe.apiml.util.EurekaUtils;
 
 import java.util.List;
 
@@ -28,13 +31,14 @@ import java.util.List;
  * and using or releasing on gateways sides the shortest.
  */
 @Component
+@Slf4j
 public class GatewayNotifier {
 
     private final ApimlLogger logger;
 
     private final RestTemplate restTemplate;
 
-    public GatewayNotifier(RestTemplate restTemplate, MessageService messageService) {
+    public GatewayNotifier(@Qualifier("restTemplateWithKeystore") RestTemplate restTemplate, MessageService messageService) {
         this.restTemplate = restTemplate;
         this.logger = ApimlLogger.of(GatewayNotifier.class, messageService);
     }
@@ -52,8 +56,9 @@ public class GatewayNotifier {
      * any modification of service instance happened.
      *
      * @param serviceId id of modified service
+     * @param instanceId is of modified instance, if it is available
      */
-    public void serviceUpdated(String serviceId) {
+    public void serviceUpdated(String serviceId, String instanceId) {
         final PeerAwareInstanceRegistry registry = getRegistry();
         final Application application = registry.getApplication("GATEWAY");
         if (application == null) {
@@ -64,11 +69,19 @@ public class GatewayNotifier {
         final List<InstanceInfo> gatewayInstances = application.getInstances();
 
         for (final InstanceInfo instanceInfo : gatewayInstances) {
+            // don't notify service itself, it is not required
+            if (StringUtils.equalsIgnoreCase(instanceId, instanceInfo.getInstanceId())) continue;
+
             final StringBuilder url = new StringBuilder();
             url.append(EurekaUtils.getUrl(instanceInfo)).append("/cache/services");
-            if (serviceId != null)
-                url.append('/').append(serviceId);
-            restTemplate.delete(url.toString());
+            if (serviceId != null) url.append('/').append(serviceId);
+
+            try {
+                restTemplate.delete(url.toString());
+            } catch (Exception e) {
+                log.debug("Cannot notify the Gateway {} about {}", url.toString(), instanceId, e);
+                logger.log("org.zowe.apiml.discovery.registration.gateway.notify", url.toString(), instanceId);
+            }
         }
     }
